clean_UPT/the_well/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["the_well*"]

[tool.setuptools.package-data]
the_well = ["utils/registry.yaml"]

[project]
name = "the_well"
dynamic = ["version"]
description = "A large-scale collection of machine learning datasets of various spatiotemporal physical systems"
license = { file = "LICENSE" }
readme = "README.md"
authors = [
    {name="Miles Cranmer"},
    {name="Michael McCabe"},
    {name="Lucas Meyer"},
    {name="Rudy Morel"},
    {name="Ruben Ohana"},
    {name="Bruno Régaldo-Saint Blancard"},
    {name="François Rozet"},
]
maintainers = [
    {name="Miles Cranmer"},
    {name="Mike McCabe"},
    {name="Lucas Meyer"},
    {name="Payel Mukhopadhyay"},
    {name="Ruben Ohana"},
    {name="François Rozet"},
]
requires-python = ">=3.9"
dependencies = [
    "einops>=0.8",
    "h5py>=3.9.0",
    "numpy>=1.20",
    "pyyaml>=6.0",
    "torch>=2.1",
    'typing_extensions>=4.6; python_version < "3.11"',
    'eval-type-backport>=0.2; python_version < "3.10"',
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
]

[project.urls]
documentation = "https://polymathic-ai.org/the_well"
organization = "https://polymathic-ai.org/"
source = "https://github.com/PolymathicAI/the_well"
tracker = "https://github.com/PolymathicAI/the_well/issues"


[project.scripts]
the-well-download = "the_well.utils.download:main"

[project.optional-dependencies]
benchmark = [
    "hydra-core>=1.3,<2",
    "matplotlib>=3.8,<4",
    "neuraloperator==0.3.0",
    "omegaconf>=2.0,<3",
    "plotly>=5.0,<6",
    "timm>=1.0,<2",
    "torch_harmonics>=0.6,<0.7",
    "wandb>=0.17.9,<0.18",
    "torchinfo>=1.8.0,<2",
    "huggingface-hub"
]
dev = [
    "pre-commit",
    "pytest",
    "pytest-order",
    "ruff",
]
docs = [
    "mkdocs==1.6.1",
    "mkdocs-jupyter==0.25.1",
    "mkdocs-material==9.5.45",
    "mkdocstrings-python==1.12.2",
]

[tool.setuptools.dynamic]
version = {attr = "the_well.__version__"}

[tool.ruff]
extend-include = ["*.ipynb"]

[tool.ruff.format]
exclude = ["*.ipynb"]

[tool.ruff.lint]
extend-select = ["I", "W"]

[tool.ruff.lint.extend-per-file-ignores]
"datasets/**/*.py" = ["F841"]
